@@ -21,60 +21,60 @@
   },
   "dependencies": {
     "@astro-community/astro-embed-youtube": "^0.4.1",
-    "@astrojs/image": "^0.17.3",
-    "@astrojs/mdx": "^0.19.7",
-    "@astrojs/prefetch": "^0.3.0",
-    "@astrojs/rss": "^2.4.4",
-    "@astrojs/sitemap": "^2.0.2",
-    "@astrojs/solid-js": "^2.0.1",
-    "@astrojs/tailwind": "^4.0.0",
+    "@astrojs/mdx": "^1.0.1",
+    "@astrojs/prefetch": "^0.4.0",
+    "@astrojs/rss": "^3.0.0",
+    "@astrojs/sitemap": "^3.0.0",
+    "@astrojs/solid-js": "^3.0.0",
+    "@astrojs/tailwind": "^5.0.0",
     "@fontsource-variable/inter": "^5.0.8",
-    "astro": "^2.10.9",
-    "chart.js": "4.2.1",
-    "clsx": "^1.2.1",
+    "astro": "^3.0.9",
+    "chart.js": "4.4.0",
+    "clsx": "^2.0.0",
     "gray-matter": "^4.0.3",
-    "node-fetch": "^3.3.0",
+    "node-fetch": "^3.3.2",
     "quill": "^1.3.7",
     "sharp": "^0.32.5",
-    "slugify": "^1.6.5",
-    "solid-js": "^1.6.10",
+    "slugify": "^1.6.6",
+    "solid-js": "^1.7.11",
     "tailwindcss": "^3.3.3",
     "xss": "^1.0.14",
-    "zod": "^3.21.4",
-    "zod-form-data": "^1.3.1"
+    "zod": "^3.22.2",
+    "zod-form-data": "^2.0.1"
   },
   "devDependencies": {
     "@actions/core": "^1.10.0",
-    "@astrojs/site-kit": "github:withastro/site-kit#f1f73aec5fc32213bc55984af7c1033105a96cbd",
-    "@astrojs/vercel": "^3.8.2",
-    "@babel/core": "^7.20.12",
+    "@astrojs/check": "^0.2.0",
+    "@astrojs/site-kit": "github:withastro/site-kit",
+    "@astrojs/vercel": "^4.0.4",
+    "@babel/core": "^7.22.15",
     "@octokit/graphql": "^5.0.5",
-    "@playwright/test": "^1.30.0",
-    "@types/node": "^18.14.2",
-    "@types/quill": "^2.0.10",
-    "@typescript-eslint/eslint-plugin": "^5.50.0",
-    "@typescript-eslint/parser": "^5.50.0",
-    "astro-eslint-parser": "^0.11.0",
-    "astro-icon": "^0.8.0",
-    "date-fns": "^2.29.3",
-    "dotenv": "^16.0.3",
-    "eslint": "^8.33.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-astro": "^0.23.0",
+    "@playwright/test": "^1.37.1",
+    "@types/node": "^20.5.9",
+    "@types/quill": "^2.0.11",
+    "@typescript-eslint/eslint-plugin": "^6.6.0",
+    "@typescript-eslint/parser": "^6.6.0",
+    "astro-eslint-parser": "^0.15.0",
+    "astro-icon": "^0.8.1",
+    "date-fns": "^2.30.0",
+    "dotenv": "^16.3.1",
+    "eslint": "^8.48.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-astro": "^0.29.0",
     "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-tailwindcss": "^3.8.3",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-tailwindcss": "^3.13.0",
     "json-to-pretty-yaml": "^1.2.2",
-    "linkedom": "^0.14.25",
+    "linkedom": "^0.15.3",
     "npm-run-all": "^4.1.5",
     "organize-imports-cli": "^0.10.0",
-    "prettier": "^2.8.8",
-    "prettier-plugin-astro": "^0.10.0",
-    "prettier-plugin-tailwindcss": "^0.3.0",
-    "puppeteer": "^19.8.0",
-    "smartypants": "^0.1.6",
+    "prettier": "^3.0.3",
+    "prettier-plugin-astro": "^0.12.0",
+    "prettier-plugin-tailwindcss": "^0.5.4",
+    "puppeteer": "^21.1.1",
+    "smartypants": "^0.2.2",
     "tiny-glob": "^0.2.9",
-    "typescript": "^4.9.5"
+    "typescript": "^5.2.2"
   },
-  "packageManager": "pnpm@8.7.0"
+  "packageManager": "pnpm@8.7.4"
 }
