@@ -1,4 +1,4 @@
-import { JSX } from "solid-js/jsx-runtime"
+import type { JSX } from "solid-js/jsx-runtime"
 
 export default function DownloadIcon(props: JSX.SvgSVGAttributes<SVGSVGElement>) {
 	return (
